+++ src/main/java/dev/root101/clean/core/rest/CRUDRestService.java
@@ -16,28 +16,28 @@
  */
 package dev.root101.clean.core.rest;
 
-import dev.root101.clean.core.app.usecase.AbstractUseCase;
+import dev.root101.clean.core.app.domain.DomainObject;
 import java.util.List;
 
 /**
  *
  * @author Root101 (jhernandezb96@gmail.com, +53-5-426-8660)
  * @author JesusHdezWaterloo@Github
- * @param <T>
+ * @param <Domain>
  */
-public interface CRUDRestService<T> extends AbstractUseCase {
+public interface CRUDRestService<Domain extends DomainObject> extends AbstractRestService {
 
-    public T create(T newObject) throws RuntimeException;
+    public Domain create(Domain newObject) throws RuntimeException;
 
-    public T edit(T objectToEdit) throws RuntimeException;
+    public Domain edit(Domain objectToEdit) throws RuntimeException;
 
-    public T destroy(T objectToDestroy) throws RuntimeException;
+    public Domain destroy(Domain objectToDestroy) throws RuntimeException;
 
-    public T destroyById(Object keyId) throws RuntimeException;
+    public Domain destroyById(Object keyId) throws RuntimeException;
 
-    public T findBy(Object keyId) throws RuntimeException;
+    public Domain findBy(Object keyId) throws RuntimeException;
 
-    public List<T> findAll() throws RuntimeException;
+    public List<Domain> findAll() throws RuntimeException;
 
     /**
      * By default return the size of the findAll() list.